move avatar controller to core/controller
