+++ core/application.php
@@ -31,7 +31,7 @@ use OC\AppFramework\Utility\TimeFactory;
 use \OCP\AppFramework\App;
 use OC\Core\LostPassword\Controller\LostController;
 use OC\Core\User\UserController;
-use OC\Core\Avatar\AvatarController;
+use OC\Core\Controller\AvatarController;
 use \OCP\Util;
 
 /**
